Update to remove unneeded prefix
