+++ tensorflow_io/core/python/api/experimental/__init__.py
@@ -14,6 +14,6 @@
 # ==============================================================================
 """tensorflow_io.experimental"""
 
-from tensorflow_io.core.python.ops.experimental.io_dataset_ops import IODataset
+from tensorflow_io.core.python.experimental.io_dataset_ops import IODataset
 
 from tensorflow_io.core.python.api.experimental import text
