+++ tensorflow_io/core/python/api/v0/__init__.py
@@ -15,5 +15,5 @@
 """tensorflow_io"""
 
 # tensorflow_io.core.python.ops is implicitly imported (along with file system)
-from tensorflow_io.core.python.ops.v0.io_tensor import IOTensor
-from tensorflow_io.core.python.ops.v0.io_dataset import IODataset, IOStreamDataset
+from tensorflow_io.core.python.ops.io_tensor import IOTensor
+from tensorflow_io.core.python.ops.io_dataset import IODataset, IOStreamDataset
