+++ tensorflow_io/core/python/experimental/io_dataset_ops.py
@@ -19,8 +19,8 @@ from __future__ import print_function
 
 import tensorflow as tf
 
-from tensorflow_io.core.python.ops.v0 import io_dataset
-from tensorflow_io.core.python.ops.experimental import libsvm_dataset_ops
+from tensorflow_io.core.python.ops import io_dataset
+from tensorflow_io.core.python.experimental import libsvm_dataset_ops
 
 class IODataset(io_dataset.IODataset):
   """IODataset"""
