+++ tensorflow_io/core/python/ops/audio_io_tensor_ops.py
@@ -20,7 +20,7 @@ from __future__ import print_function
 import uuid
 
 import tensorflow as tf
-from tensorflow_io.core.python.ops.v0 import io_tensor_ops
+from tensorflow_io.core.python.ops import io_tensor_ops
 from tensorflow_io.core.python.ops import core_ops
 
 class _AudioIOTensorFunction(object):
