+++ tensorflow_io/core/python/ops/hdf5_dataset_ops.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""AudioDataset"""
+"""HDF5Dataset"""
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
@@ -21,32 +21,39 @@ import uuid
 
 import tensorflow as tf
 from tensorflow_io.core.python.ops import core_ops
-from tensorflow_io.core.python.ops.v0 import io_dataset_ops
+from tensorflow_io.core.python.ops import io_dataset_ops
 
-class _AudioIODatasetFunction(object):
-  def __init__(self, resource, shape, dtype):
+class _HDF5IODatasetFunction(object):
+  def __init__(self, function, resource, component, shape, dtype):
+    self._function = function
     self._resource = resource
+    self._component = component
     self._shape = tf.TensorShape([None]).concatenate(shape[1:])
     self._dtype = dtype
   def __call__(self, start, stop):
-    return core_ops.io_wav_readable_read(
+    return self._function(
         self._resource, start=start, stop=stop,
-        shape=self._shape, dtype=self._dtype)
+        component=self._component, shape=self._shape, dtype=self._dtype)
 
-class AudioIODataset(io_dataset_ops._IODataset): # pylint: disable=protected-access
-  """AudioIODataset"""
+class HDF5IODataset(io_dataset_ops._IODataset): # pylint: disable=protected-access
+  """HDF5IODataset"""
 
   def __init__(self,
                filename,
+               dataset,
                internal=True):
-    """AudioIODataset."""
-    with tf.name_scope("AudioIODataset") as scope:
-      resource = core_ops.io_wav_readable_init(
+    """HDF5IODataset."""
+    with tf.name_scope("HDF5IODataset") as scope:
+      resource, _ = core_ops.io_hdf5_readable_init(
           filename,
           container=scope,
           shared_name="%s/%s" % (filename, uuid.uuid4().hex))
-      shape, dtype, _ = core_ops.io_wav_readable_spec(resource)
-      shape = tf.TensorShape(shape.numpy())
+      shape, dtype = core_ops.io_hdf5_readable_spec(resource, dataset)
+      shape = tf.TensorShape([None if e < 0 else e for e in shape.numpy()])
       dtype = tf.as_dtype(dtype.numpy())
-      super(AudioIODataset, self).__init__(
-          _AudioIODatasetFunction(resource, shape, dtype), internal=internal)
+      capacity = 4096
+      super(HDF5IODataset, self).__init__(
+          _HDF5IODatasetFunction(
+              core_ops.io_hdf5_readable_read,
+              resource, dataset, shape, dtype),
+          capacity=capacity, internal=internal)
