+++ tensorflow_io/core/python/ops/prometheus_io_tensor_ops.py
@@ -20,8 +20,8 @@ from __future__ import print_function
 import uuid
 
 import tensorflow as tf
+from tensorflow_io.core.python.ops import io_tensor_ops
 from tensorflow_io.core.python.ops import golang_ops
-from tensorflow_io.core.python.ops.v0 import io_tensor_ops
 
 class _PrometheusIOTensorFunction(object):
   """_AudioIOTensorFunction"""
