+++ tensorflow_io/json/python/ops/json_ops.py
@@ -18,7 +18,7 @@ from __future__ import division
 from __future__ import print_function
 
 import tensorflow as tf
-from tensorflow_io.core.python.ops.v0 import data_ops
+from tensorflow_io.core.python.ops import data_ops
 from tensorflow_io.core.python.ops import core_ops
 
 def list_json_columns(filename):
