+++ tensorflow_io/mnist/python/ops/mnist_ops.py
@@ -18,7 +18,7 @@ from __future__ import division
 from __future__ import print_function
 
 import tensorflow as tf
-from tensorflow_io.core.python.ops.v0 import data_ops as data_ops
+from tensorflow_io.core.python.ops import data_ops as data_ops
 from tensorflow_io.core.python.ops import core_ops as mnist_ops
 
 class MNISTLabelDataset(data_ops.Dataset):
